pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg==0.31.0",
    "fastapi==0.128.0",
    "firecrawl==4.13.4",
    "langchain==1.2.7",
    "langchain-community==0.4.1",
    "langchain-core==1.2.7",
    "langchain-huggingface==1.2.0",
    "langchain-postgres>=0.0.16",
    "langgraph==1.0.7",
    "langgraph-checkpoint-postgres==3.0.3",
    "pgvector>=0.3.6",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.4",
    "sqlalchemy==2.0.46",
    "uvicorn==0.40.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "ruff>=0.14.14",
]
